+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix #322
 - Fix missing update of the value scale upon rerendering
 - Fix absToChr and chrToAbs of the ChromInfo API
+- Fixed a nasty rendering bug in the chrom labels track
 
 ## v1.2.3